buildlit/requirements.txt
asgiref==3.9.1
asttokens==3.0.0
black==25.1.0
cfgv==3.4.0
click==8.2.1
colorama==0.4.6
coverage==7.10.2
decorator==5.2.1
distlib==0.4.0
dj-database-url==3.0.1
Django==5.2.5
django-cors-headers==4.7.0
django-debug-toolbar==6.0.0
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
executing==2.2.0
filelock==3.18.0
identify==2.6.12
inflection==0.5.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isort==6.0.1
jedi==0.19.2
Markdown==3.8.2
matplotlib-inline==0.1.7
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
parso==0.8.4
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pre_commit==4.2.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
tzdata==2025.2
uritemplate==4.2.0
virtualenv==20.33.1
wcwidth==0.2.13
